+++ VComponents/Components/Buttons/Primary Button/VPrimaryButtonModel.swift
@@ -119,6 +119,8 @@ public struct VPrimaryButtonModel {
         /// Custom content opacities.
         ///
         /// Applicable only when init with content is used.
+        /// When using a custom content, it's subviews cannot be configured with indivudual colors,
+        /// so instead, a general opacity is being applied.
         public var customContentOpacities: StateOpacities = .init(
             enabled: 1,
             pressed: 0.75,