+++ VComponents/Components/Item Pickers/Menu Picker/VMenuPicker.swift
@@ -10,19 +10,17 @@ import SwiftUI
 // MARK: - V Menu Picker
 /// Item picker component that selects from a set of mutually exclusive values, and displays their representative content in a menu.
 ///
-/// Component can be initialized with data, row titles, `PickableEnumeration`, or `PickableTitledItem`.
+/// Component row content can be initialized with data, row titles, or `PickableEnumeration`/`PickableTitledEnumeration`.
 ///
-/// Component can be initialized with content or title.
-///
-/// Component supports presets or existing button types.
+/// Component label can be used with button components from the library.
 ///
 /// Best suited for `5+` items.
 ///
-/// State can be passed as parameter.
+/// Model can be passed as parameter.
 ///
 /// Usage example:
 ///
-///     enum PickerRow: Int, PickableEnumeration {
+///     enum PickerRow: Int, PickableTitledEnumeration {
 ///         case red, green, blue
 ///
 ///         var pickerTitle: String {
@@ -32,21 +30,17 @@ import SwiftUI
 ///             case .blue: return "Blue"
 ///             }
 ///         }
-///
-///         var pickerIcon: String { "swift" }
 ///     }
 ///
 ///     @State var selection: PickerRow = .red
 ///
 ///     var body: some View {
 ///         VMenuPicker(
-///             preset: .secondary(),
 ///             selection: $selection,
-///             title: "Lorem Ipsum",
-///             rowContent: { row in
-///                 .titledSystemIcon(
-///                     title: row.pickerTitle,
-///                     name: "swift"
+///             label: {
+///                 VPlainButton(
+///                     action: {},
+///                     title: "Select"
 ///                 )
 ///             }
 ///         )
@@ -59,208 +53,66 @@ public struct VMenuPicker<Label, Data>: View
         Data.Index == Int
 {
     // MARK: Properties
-    private let menuPickerButtonType: VMenuPickerButtonType
-    
-    private let state: VMenuPickerState
+    @Environment(\.isEnabled) private var isEnabled: Bool
+    private var internalState: VMenuPickerInternalState { .init(isEnabled: isEnabled) }
     
     @Binding private var selectedIndex: Int
     
     private let label: () -> Label
     private let data: Data
-    private let rowContent: (Data.Element) -> VMenuPickerRow
+    private let rowContent: (Data.Element) -> VMenuPickerRowContent
 
-    // MARK: Initializers - View Builder and Preset
-    /// Initializes component with preset, selected index, content, data, and row content.
+    // MARK: Initializers - View Builder
+    /// Initializes component with selected index, label, data, and row content.
     public init(
-        preset menuPickerButtonPreset: VMenuPickerButtonPreset,
-        state: VMenuPickerState = .enabled,
         selectedIndex: Binding<Int>,
         @ViewBuilder label: @escaping () -> Label,
         data: Data,
-        rowContent: @escaping (Data.Element) -> VMenuPickerRow
+        rowContent: @escaping (Data.Element) -> VMenuPickerRowContent
     ) {
-        self.menuPickerButtonType = menuPickerButtonPreset.buttonType
-        self.state = state
         self._selectedIndex = selectedIndex
         self.label = label
         self.data = data
         self.rowContent = rowContent
     }
-    
-    /// Initializes component with preset, selected index, title, data, and row content.
+
+    // MARK: Initializers - Row Titles
+    /// Initializes component with selected index, label, and row titles.
     public init(
-        preset menuPickerButtonPreset: VMenuPickerButtonPreset,
-        state: VMenuPickerState = .enabled,
-        selectedIndex: Binding<Int>,
-        title: String,
-        data: Data,
-        rowContent: @escaping (Data.Element) -> VMenuPickerRow
-    )
-        where Label == VText
-    {
-        self.init(
-            preset: menuPickerButtonPreset,
-            state: state,
-            selectedIndex: selectedIndex,
-            label: { menuPickerButtonPreset.text(from: title, isEnabled: state.isEnabled) },
-            data: data,
-            rowContent: rowContent
-        )
-    }
-    
-    // MARK: Initializers - View Builder and Custom
-    /// Initializes component with selected index, content, data, and row content.
-    public init(
-        state: VMenuPickerState = .enabled,
-        selectedIndex: Binding<Int>,
-        @ViewBuilder label: @escaping () -> Label,
-        data: Data,
-        rowContent: @escaping (Data.Element) -> VMenuPickerRow
-    ) {
-        self.menuPickerButtonType = .custom
-        self.state = state
-        self._selectedIndex = selectedIndex
-        self.label = label
-        self.data = data
-        self.rowContent = rowContent
-    }
-    
-    // MARK: Initializers - Row Titles and Preset
-    /// Initializes component with preset, selected index, content, and row titles.
-    public init(
-        preset menuPickerButtonPreset: VMenuPickerButtonPreset,
-        state: VMenuPickerState = .enabled,
         selectedIndex: Binding<Int>,
         @ViewBuilder label: @escaping () -> Label,
         rowTitles: [String]
     )
         where Data == Array<String>
     {
-        self.menuPickerButtonType = menuPickerButtonPreset.buttonType
-        self.state = state
         self._selectedIndex = selectedIndex
         self.label = label
         self.data = rowTitles
-        self.rowContent = { title in .titled(title: title) }
+        self.rowContent = { .title(title: $0) }
     }
     
-    /// Initializes component with preset, selected index, title, and row titles.
-    public init(
-        preset menuPickerButtonPreset: VMenuPickerButtonPreset,
-        state: VMenuPickerState = .enabled,
-        selectedIndex: Binding<Int>,
-        title: String,
-        rowTitles: [String]
-    )
-        where
-            Label == VText,
-            Data == Array<String>
-    {
-        self.init(
-            preset: menuPickerButtonPreset,
-            state: state,
-            selectedIndex: selectedIndex,
-            label: { menuPickerButtonPreset.text(from: title, isEnabled: state.isEnabled) },
-            rowTitles: rowTitles
-        )
-    }
-    
-    // MARK: Initializers - Row Titles and Custom
-    /// Initializes component with selected index, content, and row titles.
-    public init(
-        state: VMenuPickerState = .enabled,
-        selectedIndex: Binding<Int>,
-        @ViewBuilder label: @escaping () -> Label,
-        rowTitles: [String]
-    )
-        where Data == Array<String>
-    {
-        self.init(
-            state: state,
-            selectedIndex: selectedIndex,
-            label: label,
-            data: rowTitles,
-            rowContent: { title in .titled(title: title) }
-        )
-    }
-    
-    // MARK: Initializers - Pickable Item and Preset
-    /// Initializes component with preset, `PickableEnumeration`, content, and row content.
+    // MARK: Initializers - Pickable Enumeration & Pickable Titled Enumeration
+    /// Initializes component with `PickableEnumeration`, label, and row content.
     public init<PickableItem>(
-        preset menuPickerButtonPreset: VMenuPickerButtonPreset,
-        state: VMenuPickerState = .enabled,
         selection: Binding<PickableItem>,
         @ViewBuilder label: @escaping () -> Label,
-        rowContent: @escaping (PickableItem) -> VMenuPickerRow
+        rowContent: @escaping (PickableItem) -> VMenuPickerRowContent
     )
         where
             Data == Array<PickableItem>,
             PickableItem: PickableEnumeration
     {
-        self.init(
-            preset: menuPickerButtonPreset,
-            state: state,
-            selectedIndex: .init(
-                get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
-                set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
-            ),
-            label: label,
-            data: .init(PickableItem.allCases),
-            rowContent: rowContent
+        self._selectedIndex = .init(
+            get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
+            set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
         )
+        self.label = label
+        self.data = .init(PickableItem.allCases)
+        self.rowContent = rowContent
     }
-    
-    /// Initializes component with preset, `PickableEnumeration`, title, and row content.
+
+    /// Initializes component with `PickableTitledItem` and label.
     public init<PickableItem>(
-        preset menuPickerButtonPreset: VMenuPickerButtonPreset,
-        state: VMenuPickerState = .enabled,
-        selection: Binding<PickableItem>,
-        title: String,
-        rowContent: @escaping (PickableItem) -> VMenuPickerRow
-    )
-        where
-            Label == VText,
-            Data == Array<PickableItem>,
-            PickableItem: PickableEnumeration
-    {
-        self.init(
-            preset: menuPickerButtonPreset,
-            state: state,
-            selection: selection,
-            label: { menuPickerButtonPreset.text(from: title, isEnabled: state.isEnabled) },
-            rowContent: rowContent
-        )
-    }
-    
-    // MARK: Initializers - Pickable Item and Custom
-    /// Initializes component with `PickableEnumeration`, content, and row content.
-    public init<PickableItem>(
-        state: VMenuPickerState = .enabled,
-        selection: Binding<PickableItem>,
-        @ViewBuilder label: @escaping () -> Label,
-        rowContent: @escaping (PickableItem) -> VMenuPickerRow
-    )
-        where
-            Data == Array<PickableItem>,
-            PickableItem: PickableEnumeration
-    {
-        self.init(
-            state: state,
-            selectedIndex: .init(
-                get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
-                set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
-            ),
-            label: label,
-            data: .init(PickableItem.allCases),
-            rowContent: rowContent
-        )
-    }
-    
-    // MARK: Initializers - Pickable Titled Item and Preset
-    /// Initializes component with preset, `PickableTitledItem` and content.
-    public init<PickableItem>(
-        preset menuPickerButtonPreset: VMenuPickerButtonPreset,
-        state: VMenuPickerState = .enabled,
         selection: Binding<PickableItem>,
         @ViewBuilder label: @escaping () -> Label
     )
@@ -268,125 +120,77 @@ public struct VMenuPicker<Label, Data>: View
             Data == Array<PickableItem>,
             PickableItem: PickableTitledEnumeration
     {
-        self.init(
-            preset: menuPickerButtonPreset,
-            state: state,
-            selectedIndex: .init(
-                get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
-                set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
-            ),
-            label: label,
-            data: .init(PickableItem.allCases),
-            rowContent: { item in .titled(title: item.pickerTitle) }
+        self._selectedIndex = .init(
+            get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
+            set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
         )
-    }
-    
-    /// Initializes component with preset, `PickableTitledItem` and title.
-    public init<PickableItem>(
-        preset menuPickerButtonPreset: VMenuPickerButtonPreset,
-        state: VMenuPickerState = .enabled,
-        selection: Binding<PickableItem>,
-        title: String
-    )
-        where
-            Label == VText,
-            Data == Array<PickableItem>,
-            PickableItem: PickableTitledEnumeration
-    {
-        self.init(
-            preset: menuPickerButtonPreset,
-            state: state,
-            selectedIndex: .init(
-                get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
-                set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
-            ),
-            label: { menuPickerButtonPreset.text(from: title, isEnabled: state.isEnabled) },
-            data: .init(PickableItem.allCases),
-            rowContent: { item in .titled(title: item.pickerTitle) }
-        )
-    }
-    
-    // MARK: Initializers - Pickable Titled Item and Custom
-    /// Initializes component with `PickableTitledItem` and content.
-    public init<PickableItem>(
-        state: VMenuPickerState = .enabled,
-        selection: Binding<PickableItem>,
-        @ViewBuilder label: @escaping () -> Label
-    )
-        where
-            Data == Array<PickableItem>,
-            PickableItem: PickableTitledEnumeration
-    {
-        self.init(
-            state: state,
-            selectedIndex: .init(
-                get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
-                set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
-            ),
-            label: label,
-            data: .init(PickableItem.allCases),
-            rowContent: { item in .titled(title: item.pickerTitle) }
-        )
+        self.label = label
+        self.data = .init(PickableItem.allCases)
+        self.rowContent = { .title(title: $0.pickerTitle) }
     }
 
     // MARK: Body
     public var body: some View {
         Menu(
+            content: { menuContent },
+            label: { label() }
+        )
+            .disabled(!internalState.isEnabled)
+    }
+    
+    private var menuContent: some View {
+        Picker(
+            selection: $selectedIndex,
             content: {
-                Picker(
-                    selection: $selectedIndex,
-                    content: {
-                        ForEach(data.enumeratedArray().reversed(), id: \.offset, content: { (i, row) in
-                            rowView(rowContent(row))
-                                .tag(i)
-                        })
-                    },
-                    label: { EmptyView() }
-                )
-                    .pickerStyle(.inline)
+                ForEach(data.enumeratedArray().reversed(), id: \.offset, content: { (i, row) in
+                    rowView(rowContent(row))
+                        .tag(i)
+                })
             },
-            label: { labelView }
+            label: { EmptyView() }
         )
-            .disabled(!state.isEnabled) // Luckily, doesn't affect colors
+            .pickerStyle(.inline)
     }
     
-    private var labelView: some View {
-        VMenuPickerButtonType.menuPickerButton(
-            buttonType: menuPickerButtonType,
-            isEnabled: state.isEnabled,
-            label: label
-        )
-    }
-    
-    @ViewBuilder private func rowView(_ row: VMenuPickerRow) -> some View {
+    @ViewBuilder private func rowView(_ row: VMenuPickerRowContent) -> some View {
         switch row {
-        case .titled(let title):
+        case .title(let title):
             Text(title)
             
-        case .titledSystemIcon(let title, let name):
+        case .titleIcon(let title, let icon):
             HStack(content: {
                 Text(title)
-                Image(systemName: name)
+                icon
             })
-        
-        case .titledAssetIcon(let title, let name, let bundle):
+            
+        case .titleAssetIcon(let title, let icon, let bundle):
             HStack(content: {
                 Text(title)
-                Image(name, bundle: bundle)
+                Image(icon, bundle: bundle)
+            })
+            
+        case .titleSystemIcon(let title, let icon):
+            HStack(content: {
+                Text(title)
+                Image(systemName: icon)
             })
         }
     }
 }
 
 // MARK: - Preview
-//struct VMenuPicker_Previews: PreviewProvider { // FIXME: Resolve
-//    @State private static var selection: VSegmentedPicker_Previews.PickerRow = .red
-//
-//    static var previews: some View {
-//        VMenuPicker(
-//            preset: .secondary(),
-//            selection: $selection,
-//            title: "Lorem Ipsum"
-//        )
-//    }
-//}
+struct VMenuPicker_Previews: PreviewProvider {
+    @State private static var selection: VSegmentedPicker_Previews.PickerRow = .red
+
+    static var previews: some View {
+        VMenuPicker(
+            selection: $selection,
+            label: {
+                VPlainButton(
+                    action: {},
+                    title: "Select"
+                )
+            }
+        )
+    }
+}