+++ VComponents/Components/Item Pickers/Menu Picker/VMenuPickerState.swift
@@ -8,21 +8,4 @@
 import Foundation
 
 // MARK: - V Menu Picker State
-/// Enum that describes state, such as `enabled` or `disabled`.
-public enum VMenuPickerState: Int, CaseIterable {
-    // MARK: Cases
-    /// Enabled.
-    case enabled
-    
-    /// Disabled.
-    case disabled
-    
-    // MARK: Properties
-    /// Indicates if state is enabled.
-    public var isEnabled: Bool {
-        switch self {
-        case .enabled: return true
-        case .disabled: return false
-        }
-    }
-}
+typealias VMenuPickerInternalState = GenericState_ED