+++ VComponents/Components/Item Pickers/Segmented Picker/VSegmentedPickerModel.swift
@@ -109,6 +109,8 @@ public struct VSegmentedPickerModel {
         /// Custom content opacities.
         ///
         /// Applicable only when init with content is used.
+        /// When using a custom content, it's subviews cannot be configured with indivudual colors,
+        /// so instead, a general opacity is being applied.
         public var customContentOpacities: RowStateOpacities = .init(
             enabled: 1,
             pressed: 0.5,