+++ VComponents/Components/State Pickers/Radio Button/VRadioButtonModel.swift
@@ -97,6 +97,8 @@ public struct VRadioButtonModel {
         /// Custom content opacities.
         ///
         /// Applicable only when init with content is used.
+        /// When using a custom content, it's subviews cannot be configured with indivudual colors,
+        /// so instead, a general opacity is being applied.
         public var customContentOpacities: StateOpacities = toggleRefrence.colors.customContentOpacities
 
         // MARK: Initializers