+++ VComponents/Components/State Pickers/Toggle/VToggleModel.swift
@@ -94,6 +94,8 @@ public struct VToggleModel {
         /// Custom content opacities.
         ///
         /// Applicable only when init with content is used.
+        /// When using a custom content, it's subviews cannot be configured with indivudual colors,
+        /// so instead, a general opacity is being applied.
         public var customContentOpacities: StateOpacities = .init(
             off: 1,
             on: 1,