+++ VComponentsDemo/Scenes/Demos/Item Pickers/VMenuPickerDemoView.swift
@@ -13,10 +13,9 @@ struct VMenuPickerDemoView: View {
     // MARK: Properties
     static var navBarTitle: String { "Menu Picker" }
     
-    @State private var selection: VSegmentedPickerDataSource = .red
-    @State private var state: VMenuPickerState = .enabled
-    @State private var menuPickerButtonType: VMenuPickerButtonTypeHelper = .secondary
-    @State private var contentType: VSegmentedPickerContent = .title
+    @State private var selection: VMenuPickerDataSource = .red
+    @State private var isEnabled: Bool = true
+    @State private var contentType: VMenuPickerContent = .title
 
     // MARK: Body
     var body: some View {
@@ -25,117 +24,57 @@ struct VMenuPickerDemoView: View {
     }
     
     @ViewBuilder private func component() -> some View {
-        fatalError() // FIXME: Resolve
-//        switch (menuPickerButtonType.preset, contentType) {
-//        case (let preset?, .title):
-//            VMenuPicker(
-//                preset: preset,
-//                state: state,
-//                selection: $selection,
-//                title: buttonTitle
-//            )
-//
-//        case (let preset?, .custom):
-//            VMenuPicker(
-//                preset: preset,
-//                state: state,
-//                selection: $selection,
-//                label: buttonContent
-//            )
-//
-//        case (nil, _):
-//            VMenuPicker(
-//                state: state,
-//                selection: $selection,
-//                label: buttonContent
-//            )
-//        }
+        Group(content: {
+            switch contentType {
+            case .title:
+                VMenuPicker(
+                    selection: $selection,
+                    label: {
+                        VPlainButton(
+                            action: {},
+                            title: "Lorem Ipsum"
+                        )
+                    }
+                )
+            
+            case .custom:
+                VMenuPicker(
+                    selection: $selection,
+                    label: {
+                        VPlainButton(
+                            action: {},
+                            title: "Lorem Ipsum"
+                        )
+                    },
+                    rowContent: { .titleIcon(title: $0.pickerTitle, icon: .init(systemName: "swift")) }
+                )
+            }
+        })
+            .disabled(!isEnabled)
     }
     
-    @DemoViewSettingsSectionBuilder private func settings() -> some View {
-        DemoViewSettingsSection(content: {
-            VSegmentedPicker(selection: $state, headerTitle: "State")
-        })
+    @ViewBuilder private func settings() -> some View {
+        VSegmentedPicker(
+            selection: .init(
+                get: { VMenuPickerState(isEnabled: isEnabled) },
+                set: { isEnabled = $0 == .enabled }
+            ),
+            headerTitle: "State"
+        )
 
-        DemoViewSettingsSection(content: {
-            VWheelPicker(
-                selection: $menuPickerButtonType,
-                headerTitle: "Type"
-            )
-                .onChange(of: menuPickerButtonType, perform: { type in
-                    if type == .custom { contentType = .custom }
-                })
-
-//            VSegmentedPicker( // FIXME: Resolve
-//                selection: $contentType,
-//                headerTitle: "Content",
-//                disabledItems: menuPickerButtonType == .custom ? [.title] : []
-//            )
-        })
+        VSegmentedPicker(
+            selection: $contentType,
+            headerTitle: "Content"
+        )
     }
-    
-    private var buttonTitle: String {
-        fatalError() // FIXME: Resolve
-        
-//        switch menuPickerButtonType.preset {
-//        case .square: return "Lorem"
-//        default: return "Lorem Ipsum"
-//        }
-    }
-
-    private func buttonContent() -> some View { DemoIconContentView() }
 }
 
 // MARK: - Helpers
-extension VMenuPickerState: PickableTitledEnumeration {
-    public var pickerTitle: String {
-        switch self {
-        case .enabled: return "Enabled"
-        case .disabled: return "Disabled"
-        @unknown default: fatalError()
-        }
-    }
-}
+private typealias VMenuPickerState = VSecondaryButtonState
 
-private enum VMenuPickerButtonTypeHelper: Int, PickableTitledEnumeration {
-    case primary
-    case secondary
-    case square
-    case plain
-    case custom
-    
-//    var preset: VWebLinkPreset? {
-//        switch self {
-//        case .primary: return .primary()
-//        case .secondary: return .secondary()
-//        case .square: return .square()
-//        case .plain: return .plain()
-//        case .custom: return nil
-//        }
-//    }
-    
-    var pickerTitle: String {
-        switch self {
-        case .primary: return "Primary"
-        case .secondary: return "Secondary"
-        case .square: return "Square"
-        case .plain: return "Plain"
-        case .custom: return "Custom"
-        }
-    }
-}
+private typealias VMenuPickerContent = VSegmentedPickerContent
 
-extension VMenuButtonPreset {
-    fileprivate var helperType: VMenuPickerButtonTypeHelper {
-        switch self {
-        case .primary: return .primary
-        case .secondary: return .secondary
-        case .square: return .square
-        case .plain: return .plain
-        @unknown default: fatalError()
-        }
-    }
-}
+private typealias VMenuPickerDataSource = VSegmentedPickerDataSource
 
 // MARK: - Preview
 struct VMenuPickerDemoView_Previews: PreviewProvider {